+++ example/lib/widgets/output_console.dart
@@ -75,4 +75,3 @@ class OutputConsole extends StatelessWidget {
     );
   }
 }
-